refactored machines module
